requirements.txt
tensorflow==2.8.0
python-telegram-bot==13.3
python==3.9.10
numpy==1.24.2
opencv-python==4.7.0.68
keras==2.8.0
Keras-Preprocessing==1.1.2
copy
time